+++ src/main/java/org/trebol/jpa/services/CustomersJpaServiceImpl.java
@@ -85,7 +85,14 @@ public class CustomersJpaServiceImpl
           case "id":
             return predicate.and(qCustomer.id.eq(Long.valueOf(stringValue))); // id matching is final
           case "name":
-            predicate.and(qCustomer.person.name.eq(stringValue));
+            predicate.and(qCustomer.person.firstName.eq(stringValue)
+                    .or(qCustomer.person.lastName.eq(stringValue)));
+            break;
+          case "firstName":
+            predicate.and(qCustomer.person.firstName.eq(stringValue));
+            break;
+          case "lastName":
+            predicate.and(qCustomer.person.lastName.eq(stringValue));
             break;
           case "idNumber":
             predicate.and(qCustomer.person.idNumber.eq(stringValue));
@@ -94,7 +101,14 @@ public class CustomersJpaServiceImpl
             predicate.and(qCustomer.person.email.eq(stringValue));
             break;
           case "nameLike":
-            predicate.and(qCustomer.person.name.likeIgnoreCase("%" + stringValue + "%"));
+            predicate.and(qCustomer.person.firstName.likeIgnoreCase("%" + stringValue + "%")
+                    .or(qCustomer.person.lastName.likeIgnoreCase("%" + stringValue + "%")));
+            break;
+          case "firstNameLike":
+            predicate.and(qCustomer.person.firstName.likeIgnoreCase("%" + stringValue + "%"));
+            break;
+          case "lastNameLike":
+            predicate.and(qCustomer.person.lastName.likeIgnoreCase("%" + stringValue + "%"));
             break;
           case "idNumberLike":
             predicate.and(qCustomer.person.idNumber.likeIgnoreCase("%" + stringValue + "%"));