+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Removed request authentication for remapped paths in `/data/product_categories/*` to make them publicly available to anonymous users
+- Updated `Person` schema by splitting `name` property into two properties `firstName` and `lastName`
 
 ### Deprecated
 - Path `/public/categories` and `/public/categories/{parentId}`, mapped under `PublicCategoriesController` - Thanks @ParundeepSingh